fix: created objects

+++ src/settings.rs
@@ -1,4 +1,4 @@
-pub const TILE_SIZE: u32 = 64;
+pub const TILE_SIZE: f32 = 64.0;
 pub const SCREEN_WIDTH: u32 = 1280;
 pub const SCREEN_HEIGHT: u32 = 720;
 pub const CAMERA_SCALE: f32 = 2.;
